inbox [nfc]: Move AtMentionMarker to unread_count_badge
AtMentionMarker is going to be used in subscription_list as well as
inbox, so it would be better to reuse it by having it declared as a
public class.
